Domains: Fix domain suggestions layout for tablet resolutions (#55465)
Changes proposed in this Pull Request:

This PR fixes a layout bug introduced by #55034 - domain suggestions were misaligned when viewed in tablet resolutions (> 480px and < 1080px). It affected both the domain suggestions list in the "Add a domain to this site" page and in the signup flow. Reported in p1628779410064200-slack-C020919TDRA.

Testing instructions:

Open the live Calypso link, change your browser resolution to "iPad" (768 x 1024 px), go to the /start signup flow, search for domains and ensure the domain suggestions are not misaligned.

Commits:

* Fix domain suggestions layout for tablet resolutions

* Fix pricing and "Select" button alignment in tablet resolution